Add Storybook stories for WorkflowViewer component
- Created basic stories for WorkflowViewer including HelloWorld, ComplexWorkflow, WithErrors, Interactive, MinimalView, and ReadOnlyPresentationMode.
- Added examples showcasing real-world workflows such as ECommerceWorkflow, DataProcessingPipeline, and YAML format support.
- Implemented feature stories demonstrating MiniMap, draggable nodes, multi-select functionality, and keyboard navigation.
- Introduced loading stories for workflows from URLs and file uploads, including dark theme variations.
- Developed theme stories to showcase light, dark, high contrast, soft, and custom themes.
- Organized all stories and workflow definitions into a dedicated index file for better structure and accessibility.
